mcp-server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rapid-mcp-server"
version = "0.1.0"
description = "Model Context Protocol server for the rapid data platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "rapid-sdk>=0.1.15",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
]

[project.scripts]
rapid-mcp-server = "rapid_mcp_server.server:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
